+++ redditrepostsleuth/adminsvc/adminsvc.py
@@ -8,7 +8,8 @@ sys.path.append('./')
 from redditrepostsleuth.core.notification.notification_service import NotificationService
 from redditrepostsleuth.adminsvc.misc_admin_tasks import remove_expired_bans, update_banned_sub_wiki, \
     send_reports_to_meme_voting, update_top_image_reposts, \
-    update_monitored_sub_data, check_meme_template_potential_votes, update_ban_list, queue_config_updates
+    update_monitored_sub_data, check_meme_template_potential_votes, update_ban_list, queue_config_updates, \
+    queue_post_watch_cleanup
 from redditrepostsleuth.adminsvc.inbox_monitor import InboxMonitor
 from redditrepostsleuth.core.services.eventlogging import EventLogging
 from redditrepostsleuth.core.services.response_handler import ResponseHandler
@@ -131,6 +132,14 @@ if __name__ == '__main__':
         trigger='interval',
         minutes=2,
         name='queue_config_update_check',
+        max_instances=1
+    )
+    scheduler.add_job(
+        func=queue_post_watch_cleanup,
+        args=(uowm, config),
+        trigger='interval',
+        days=3,
+        name='queue_deleted_watch_check',
         max_instances=1
     )
     scheduler.start()