+++ redditrepostsleuth/adminsvc/misc_admin_tasks.py
@@ -260,4 +260,4 @@ if __name__ == '__main__':
     notification_svc = NotificationService(config)
     reddit = get_reddit_instance(config)
     uowm = SqlAlchemyUnitOfWorkManager(get_db_engine(config))
-    queue_post_watch_cleanup(uowm)
+    update_monitored_sub_data(uowm)